# laravel13_pest_mysql/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apple Inc. API Management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/productlist/{page}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated list of products",
                "operationId": "733b41547ba73f815fd1808de7c80104",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "The page number to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product Retrieved Successfully."
                                        },
                                        "totalrecords": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "totpage": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Products is empty"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/products/search/{key}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search products by description",
                "operationId": "c35632099a120dd1b7e88d175c3fcd50",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Search keyword for descriptions",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Searched found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Searched found.."
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
        "/api/profileupdate/{id}": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user details",
                "operationId": "b94a8edf787a8c433eecbfb70897c407",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "401": {
                        "description": "Un-Authorized Access"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/uploadpicture/{id}": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update profile picture",
                "operationId": "4cfa476519579200796d8b57a223e1e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "profilepic"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "profilepic": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Picture uploaded"
                    },
                    "404": {
                        "description": "User or Image not found"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout current user",
                "description": "Deletes the current access token",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Logged out successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product Model",
                "description": "Product data structure",
                "type": "array",
                "items": {
                    "properties": {
                        "id": {
                            "type": "integer",
                            "example": 1
                        },
                        "category": {
                            "type": "string",
                            "example": "Electronics"
                        },
                        "descriptions": {
                            "type": "string",
                            "example": "Gaming Laptop"
                        },
                        "qty": {
                            "type": "integer",
                            "example": 10
                        },
                        "unit": {
                            "type": "string",
                            "example": "pcs"
                        },
                        "costprice": {
                            "type": "number",
                            "format": "float",
                            "example": 800
                        },
                        "sellprice": {
                            "type": "number",
                            "format": "float",
                            "example": 1000
                        },
                        "saleprice": {
                            "type": "number",
                            "format": "float",
                            "example": 950
                        },
                        "productpicture": {
                            "type": "string",
                            "example": "laptop.jpg"
                        },
                        "alertstocks": {
                            "type": "integer",
                            "example": 5
                        },
                        "criticalstocks": {
                            "type": "integer",
                            "example": 2
                        },
                        "createdat": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "updatedat": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    "type": "object"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}